* Remove unnecessary SPI for MetadataPublisher
