+++ dubbo-common/src/main/java/org/apache/dubbo/metadata/MetadataPublisher.java
@@ -17,14 +17,12 @@
 package org.apache.dubbo.metadata;
 
 import org.apache.dubbo.common.URL;
-import org.apache.dubbo.common.extension.SPI;
 import org.apache.dubbo.rpc.model.ApplicationModel;
 import org.apache.dubbo.rpc.model.ServiceDescriptor;
 
 /**
  * Metadata publisher.
  */
-@SPI("default")
 public interface MetadataPublisher {
 
     /**
