--- dubbo-common/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.common.deploy.ApplicationDeployListener
@@ -1,1 +1,0 @@
-metadata-register=org.apache.dubbo.metadata.MetadataPublisherRegister
