+++ dubbo-metadata/dubbo-metadata-api/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.common.deploy.ApplicationDeployListener
@@ -1,0 +1,1 @@
+metadata-register=org.apache.dubbo.metadata.report.DefaultMetadataPublisher
